+++ src/test/java/Tests/APITests/ConsentManager/CreatePINAPITest.java
@@ -7,7 +7,6 @@ import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
 import org.testng.Assert;
-import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -44,18 +43,5 @@ public class CreatePINAPITest {
             Assert.assertEquals(response.getStatusCode(), 204);
         }
     }
-
-    @AfterTest
-    public void logoutAPI() {
-        RequestSpecification request = RestAssured.given();
-
-        request.header("Content-Type", "application/json");
-        request.header("Authorization", authToken);
-        String logoutRequestBody = new LoginUser().getCMLogoutRequestBody(new LoginUser().getCMRefreshToken());
-        request.body(logoutRequestBody);
-        Response response = request.post("/logout");
-        Assert.assertEquals(response.getStatusCode(), 200);
-    }
-
 
 }
